+++ src/modules/display-meals.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import likeMeal from './add-likes.js';
 import heart from '..';
-import displayCommentPop from './comments-popup.js';
+import displayCommentPop from './popupComment.js';
 import fetchLikes from './display-likes.js';
 import count from './meal-count.js';
 import errorMessage from './error-message.js';
